requirements.txt
annotated-types==0.7.0
asgiref==3.9.2
Django==5.2.7
django-ninja==1.4.3
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
